refactor: change all responses to custom responses
