+++ internal/adapters/api/handler/role.go
@@ -1,9 +1,11 @@
-package api
+package handler
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/iBoBoTi/project_boiler_plate/internal/adapters/api/response"
 	"github.com/iBoBoTi/project_boiler_plate/internal/core/domain"
 	"github.com/iBoBoTi/project_boiler_plate/internal/core/ports"
+	"net/http"
 )
 
 type roleHandler struct {
@@ -23,45 +25,40 @@ func NewRoleHandler(roleService ports.RoleService, permissionService ports.Permi
 func (h *roleHandler) CreateRole(c *gin.Context) {
 	role := domain.Role{}
 	if err := c.ShouldBindJSON(&role); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_request_body", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-
 	if err := h.roleService.CreateRole(&role); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-
-	c.JSON(200, gin.H{"message": "Role created successfully"})
+	response.JSON(c, "role created successfully", http.StatusCreated, nil, nil)
 }
 
 func (h *roleHandler) GetRole(c *gin.Context) {
 	id := c.Param("id")
 	role, err := h.roleService.GetRoleByID(id)
 	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-
-	c.JSON(200, gin.H{"role": role})
+	response.JSON(c, "role gotten", http.StatusCreated, role, nil)
 }
 
 func (h *roleHandler) GetRoles(c *gin.Context) {
 	roles, err := h.roleService.GetAllRoles()
 	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-
-	c.JSON(200, gin.H{"roles": roles})
+	response.JSON(c, "roles gotten", http.StatusCreated, roles, nil)
 }
 
 func (h *roleHandler) DeleteRole(c *gin.Context) {
 	id := c.Param("id")
 	if err := h.roleService.DeleteRole(id); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-
-	c.JSON(200, gin.H{"message": "Role deleted successfully"})
+	response.JSON(c, "roles deleted successfully", http.StatusCreated, nil, nil)
 }
