+++ internal/adapters/api/handler/role_permissions.go
@@ -1,9 +1,11 @@
-package api
+package handler
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/iBoBoTi/project_boiler_plate/internal/adapters/api/response"
 	"github.com/iBoBoTi/project_boiler_plate/internal/core/domain"
 	"github.com/iBoBoTi/project_boiler_plate/internal/core/ports"
+	"net/http"
 )
 
 type rolePermissionsHandler struct {
@@ -21,33 +23,33 @@ func NewRolePermissionsHandler(rolePermissionsService ports.RolePermissionsServi
 func (h *rolePermissionsHandler) AddPermissionsToRole(c *gin.Context) {
 	rolePermissions := domain.RolePermissions{}
 	if err := c.ShouldBindJSON(&rolePermissions); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_request_body", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
 	if err := h.rolePermissionsService.AddPermissionsToRole(&rolePermissions); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, gin.H{"message": "success"})
+	response.JSON(c, "permissions added successfully", http.StatusOK, nil, nil)
 }
 
 func (h *rolePermissionsHandler) RemovePermissionFromRole(c *gin.Context) {
 	roleID := c.Param("role_id")
 	permissionID := c.Param("permission_id")
 	if err := h.rolePermissionsService.RemovePermissionFromRole(roleID, permissionID); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, gin.H{"message": "success"})
+	response.JSON(c, "permission removed successfully", http.StatusOK, nil, nil)
 }
 
 func (h *rolePermissionsHandler) GetAllPermissionsForRole(c *gin.Context) {
 	roleID := c.Param("role_id")
 	permissions, err := h.rolePermissionsService.GetAllPermissionsForRole(roleID)
 	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, permissions)
+	response.JSON(c, "permissions gotten", http.StatusOK, permissions, nil)
 
 }
