+++ internal/adapters/api/handler/user.go
@@ -1,4 +1,4 @@
-package api
+package handler
 
 import (
 	"github.com/gin-gonic/gin"
